strip trailing blank lines for some rules using audit-file_contents template
There are two cases of such rules.
One group must not include the blank line at the end of the "contents" variable.
This is done by usage of "|-" Jinja marker, which automatically strips trailing blank lines from the Jinja variable.
Rules started failing because there was additional yaml content (fix text in this case) added after the definition of rule template.
The other group requires ONE blank line at the end, this case is handled in the next commit.